Migrate test helpers from main library to test target in a helpers subfolder.
