+++ clients/cpp/Client/library/tests/CMakeLists.txt
@@ -3,6 +3,10 @@ add_executable(Client-tests)
 
 set(CLIENT_TESTS_PRIVATE_SOURCES
    ${CMAKE_CURRENT_SOURCE_DIR}/sources/ClientTests.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/sources/helpers/CarpaccioStream.hpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/sources/helpers/CarpaccioStream.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/sources/helpers/TestServerResponse.hpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/sources/helpers/TestServerResponse.cpp
 )
 
 target_sources(Client-tests
@@ -14,6 +18,11 @@ source_group(
    TREE ${CMAKE_CURRENT_SOURCE_DIR}
    FILES
       ${CLIENT_TESTS_PRIVATE_SOURCES}
+)
+
+target_include_directories(Client-tests
+   PRIVATE
+      ${CMAKE_CURRENT_SOURCE_DIR}/sources
 )
 
 target_link_libraries(Client-tests
