+++ clients/cpp/Client/library/tests/sources/ClientTests.cpp
@@ -1,23 +1,22 @@
+
+#include <helpers/TestServerResponse.hpp>
+
 #include <extreme_carpaccio/client/Client.hpp>
 
 #include <extreme_carpaccio/order_management/OrderParsing.hpp>
-
-#include <gmock/gmock.h>
-#include <gtest/gtest.h>
-
-#include <iostream>
-#include <vector>
 
 #include <boost/asio/connect.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/beast/core.hpp>
 #include <boost/beast/http.hpp>
 #include <boost/beast/version.hpp>
+
 #include <cstdlib>
 #include <iostream>
 #include <string>
-#include <thread>
-#include <nlohmann/json.hpp>
+
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
 
 namespace extreme_carpaccio {
 namespace client {
@@ -30,43 +29,12 @@ using tcp = net::ip::tcp;
 
 using testing::ElementsAre;
 
-namespace {
-
-const char serverHost[] = "localhost";
-const unsigned short serverPort = 8081;
-
-static http::response<http::dynamic_body> generateServerResponse(boost::beast::http::verb requestType, const std::string & target, const std::string & contentType = "", const std::string & body = "")
-{
-   CarpaccioServer server(8081);
-   std::thread thread(&CarpaccioServer::start, &server);
-   std::this_thread::sleep_for(std::chrono::seconds(1));
-
-   CarpaccioStream stream(serverHost, serverPort);
-
-   // Send the HTTP request to the remote host
-   stream.write(requestType, target, contentType, body);
-
-   // This buffer is used for reading and must be persisted
-   beast::flat_buffer buffer;
-
-   // Receive the HTTP response
-   auto res = stream.read(buffer);
-
-   server.stop();
-
-   thread.detach();
-
-   return res;
-}
-
-} // namespace
-
 TEST(ExtremeCarpaccioClient, should_answer_404_to_incorrect_request)
 {
    const std::string target = "/je-suis-trop-fort";
 
    // Receive the HTTP response
-   const auto res = generateServerResponse(http::verb::get, target);
+   const auto res = helpers::generateServerResponse(http::verb::get, target);
    
    // Write the message to standard out
    std::cout << "Response" << std::endl << res << std::endl;
@@ -85,7 +53,7 @@ TEST(ExtremeCarpaccioClient, should_handle_post_feedback)
    std::stringstream buffer;
    std::cout.rdbuf(buffer.rdbuf());
 
-   const auto res = generateServerResponse(http::verb::post, target, "application/json", "{\"type\": \"ERROR\", \"content\": \"Field total in response is missing\"}");
+   const auto res = helpers::generateServerResponse(http::verb::post, target, "application/json", "{\"type\": \"ERROR\", \"content\": \"Field total in response is missing\"}");
 
    EXPECT_EQ(http::status::ok, res.result());
    EXPECT_EQ("ERROR : Field total in response is missing\n", buffer.str());
@@ -105,7 +73,7 @@ TEST(ExtremeCarpaccioClient, should_trace_order_description)
 	std::stringstream buffer;
 	std::cout.rdbuf(buffer.rdbuf());
 	
-	const auto res = generateServerResponse(http::verb::post, target, "application/json", "{\"prices\": [1,1.5], \"quantities\": [100, 200], \"country\": \"DE\", \"reduction\": \"STANDARD\"}");
+	const auto res = helpers::generateServerResponse(http::verb::post, target, "application/json", "{\"prices\": [1,1.5], \"quantities\": [100, 200], \"country\": \"DE\", \"reduction\": \"STANDARD\"}");
 	
    EXPECT_EQ("Order received: Order{prices={1,1.5}, quantities={100,200}, country='DE', reduction='STANDARD'}\n", buffer.str());
